# fix perf_sdt_probe to add sleep