@disgoel 2023-07-26
[debug.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/12171585/debug.log)